move mock and a test util
